+++ src/ZodBuilder/file.ts
@@ -1,4 +1,4 @@
-import type { ZodFile } from 'zod';
+import type { z, ZodFile } from 'zod';
 import type { BuilderFor } from '../Builder/index.js';
 import { ZodBuilder } from './BaseBuilder.js';
 
@@ -7,10 +7,14 @@ import { ZodBuilder } from './BaseBuilder.js';
  * Validates file uploads
  */
 export class FileBuilder
-	extends ZodBuilder<ZodFile>
+	extends ZodBuilder<ZodFile, 'file', [params?: Parameters<typeof z.file>[0]]>
 	implements BuilderFor<ZodFile>
 {
 	readonly typeKind = 'file' as const;
+
+	constructor(version?: 'v3' | 'v4', params?: Parameters<typeof z.file>[0]) {
+		super(version, params);
+	}
 
 	mime(
 		_types: string | string[],