+++ src/ZodBuilder/hostname.ts
@@ -1,4 +1,4 @@
-import type { ZodCustomStringFormat } from 'zod';
+import type { z, ZodCustomStringFormat } from 'zod';
 import type { BuilderFor } from '../Builder/index.js';
 import { StringFormatBuilder } from './StringFormatBuilder.js';
 
@@ -11,7 +11,10 @@ export class HostnameBuilder
 {
 	readonly typeKind = 'hostname' as const;
 
-	constructor(version?: 'v3' | 'v4') {
-		super(version);
+	constructor(
+		version?: 'v3' | 'v4',
+		params?: Parameters<typeof z.hostname>[0],
+	) {
+		super(version, params);
 	}
 }