+++ src/ZodBuilder/jwt.ts
@@ -1,4 +1,4 @@
-import type { ZodJWT } from 'zod';
+import type { z, ZodJWT } from 'zod';
 import type { BuilderFor } from '../Builder/index.js';
 import { StringFormatBuilder } from './StringFormatBuilder.js';
 
@@ -11,7 +11,7 @@ export class JwtBuilder
 {
 	readonly typeKind = 'jwt' as const;
 
-	constructor(version?: 'v3' | 'v4') {
-		super(version);
+	constructor(version?: 'v3' | 'v4', params?: Parameters<typeof z.jwt>[0]) {
+		super(version, params);
 	}
 }