+++ src/ZodBuilder/ksuid.ts
@@ -1,4 +1,4 @@
-import type { ZodKSUID } from 'zod';
+import type { z, ZodKSUID } from 'zod';
 import type { BuilderFor } from '../Builder/index.js';
 import { StringFormatBuilder } from './StringFormatBuilder.js';
 
@@ -11,7 +11,7 @@ export class KsuidBuilder
 {
 	readonly typeKind = 'ksuid' as const;
 
-	constructor(version?: 'v3' | 'v4') {
-		super(version);
+	constructor(version?: 'v3' | 'v4', params?: Parameters<typeof z.ksuid>[0]) {
+		super(version, params);
 	}
 }