+++ src/ZodBuilder/xid.ts
@@ -1,4 +1,4 @@
-import type { ZodXID } from 'zod';
+import type { z, ZodXID } from 'zod';
 import type { BuilderFor } from '../Builder/index.js';
 import { StringFormatBuilder } from './StringFormatBuilder.js';
 
@@ -11,7 +11,7 @@ export class XidBuilder
 {
 	readonly typeKind = 'xid' as const;
 
-	constructor(version?: 'v3' | 'v4') {
-		super(version);
+	constructor(version?: 'v3' | 'v4', params?: Parameters<typeof z.xid>[0]) {
+		super(version, params);
 	}
 }